Convert BooleanSplit to use float type as the charged field in ReconstructedParticle

+++ src/algorithms/meta/SubDivideFunctors.h
@@ -136,16 +136,16 @@ private:
 template <auto... MemberFunctionPtrs>
 class BooleanSplit {
 public:
-    using ComparisonFunction = std::function<bool(int, int)>;
+    using ComparisonFunction = std::function<bool(float, float)>;
 
-    BooleanSplit( std::vector<std::vector<int>> ids, ComparisonFunction comparison)
+    BooleanSplit( std::vector<std::vector<float>> ids, ComparisonFunction comparison)
         : m_ids(ids), m_comparisons(ids.size(), comparison) {};
 
-    BooleanSplit( std::vector<int> ids, ComparisonFunction comparison)
+    BooleanSplit( std::vector<float> ids, ComparisonFunction comparison)
         : m_ids(1,ids), m_comparisons(ids.size(), comparison) {};
 
 
-    BooleanSplit( std::vector<std::vector<int>> ids, std::vector<ComparisonFunction> comparisons)
+    BooleanSplit( std::vector<std::vector<float>> ids, std::vector<ComparisonFunction> comparisons)
         : m_ids(ids) {
         if (ids.size() != comparisons.size()) {
             throw std::invalid_argument("Size of values to compare must match the size of boolean functions");
@@ -157,7 +157,7 @@ public:
     std::vector<int> operator()(T& instance) const {
         std::vector<int> ids;
         // Check if requested value matches any configuration combinations
-        std::vector<int> values;
+        std::vector<float> values;
         (values.push_back((instance.*MemberFunctionPtrs)()), ...);
         for (size_t i = 0; i < m_ids.size(); ++i){
             if(compareVectors(m_ids[i], values, m_comparisons)){
@@ -168,10 +168,10 @@ public:
     }
 
 private:
-    std::vector<std::vector<int>> m_ids;
+    std::vector<std::vector<float>> m_ids;
     std::vector<ComparisonFunction> m_comparisons;
     
-    static bool compareVectors(const std::vector<int>& vec1, const std::vector<int>& vec2, const std::vector<ComparisonFunction>& comparisons) {
+    static bool compareVectors(const std::vector<float>& vec1, const std::vector<float>& vec2, const std::vector<ComparisonFunction>& comparisons) {
         for (size_t i = 0; i < vec1.size(); ++i) {
             if (!comparisons[i](vec1[i], vec2[i])) {
                 return false;
